OpenAI Like and Ollama UI change
From issue: https://github.com/coleam00/bolt.new-any-llm/issues/461

When selecting OpenAI Like I would like to have a a textbox for the url and textbox for model name. Since this will not be retrieved automatically. Then The Ollama provider could extend from that with a populated url but editable, and a model dropdown for the model name as it is now.